requirements.txt
torch==2.10.0
diffusers==0.36.0
transformers==5.0.0
accelerate==1.12.0
safetensors==0.7.0
pillow==12.1.0
insightface==0.7.3
onnxruntime==1.23.2
opencv-python==4.13.0.90
fastapi
uvicorn
gradio
python-multipart
